requirements.txt
openai
langchain
tiktoken
PyPDF2
qdrant_client
PyCryptodome
streamlit
htmlTemplate
InstructorEmbedding
sentence-transformers
faiss-cpu
python-dotenv
firebase_admin
st-paywall
